Strange hellos are not my bag
Added PIA (Private Internet Access) -- sorta. 

They don't share their endpoints/exits anymore (supposedly due to various streaming services using the info to block them .. killing off geo-avoidance for NetFlix/etc).

Quick tests showed that a bunch of vpn locations (very unscientific - tested 5 w/in the USA) have exits on the same /24 as the VPN concentrator IP.   So this grabs the concentrator IPs, and blocks the /24 it lives on.

This is primarily for avoiding Blueteam poking at a c2/server.